+++ test/plugin/backup_test.rb
@@ -52,6 +52,7 @@ describe "the backup plugin" do
 
   it "makes delete work with backgrounding" do
     @attacher.shrine_class.plugin :backgrounding
+    @attacher.class.stubs(:find_record)
     @attacher.class.delete { |data| (@f ||= []) << Fiber.new{self.class.delete(data)} }
     @attacher.assign(fakeio)
     replaced = @attacher._promote